Add cronJob envTests
This patch improves the existing envTest bucket to include cronJob
verification and update. It shows how an update in the Glance top level
CR is reflected in an update of the CronJob schedule. In addition, the
cronJob name now uses instance.Name instead of hardcoding the
ServiceName (if the "unique CR" name feature is used it will reflect the
name of the global CR).

Signed-off-by: Francesco Pantano <fpantano@redhat.com>