Do not create keystone endpoints when Glance is not available
Unlike services like CinderVolume, where replicas:0 still sees an
available api and scheduler, if the default GlanceAPI has replicas:0,
- which is something suggested at day1 when the storage backend is still
not available -, a keystone endpoint is created anyway in the catalog.
That means that the glance service is discovered, the endpoint can be
reached from an OpenShift point of view, but it fails returning a 503 if
a request is made.
This behavior might result confusing for the human operator, who thinks
that there's a problem with the service, ignoring that is should still
be configured.
This patch fixes this problem by **not** creating the endpoints at all
if no Glance replica is available.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>