Better handling of Topology Finalizers
When a GlanceAPI references a Topology, a finalizer is added to such
resource. By doing this, we can clearly identify which topology is
applied to which API. For this reason, and because Glance is potentially
made by a list of APIs that can independently reference the same
Topology CR, the finalizer string is based on the concatenation between
the Glance service (prefix), and the API name. It is possible, as a day2
operation, to add, update or remove a Topology reference. When doing
this, especially for update and removal operations, a pointer to the CR
must be updated. To properly support this action, the information about
the LastAppliedTopology is saved in the Status, and it allows to
properly clean up unused resources and not locking them indefinitely.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>