Fix cinder dependency when unique_pod_names is true
When unique_pod_names parameter, injected by openstack-operator is true,
the Cinder CR is created with a pseudorandom name, and it's not
predictable by glance-operator when a cinder backend is added. Other
than making sure that Cinder exists, we're not doing any other check in
glance (it's a decision), hence we can move from r.Get to r.List to
ensure that at least a Cinder CR exists in the same namespace where
Glance is deployed.

Jira: https://issues.redhat.com/browse/OSPRH-14637

Signed-off-by: Francesco Pantano <fpantano@redhat.com>