Add envTests for Cinder backend dependency
As a follow up of [1], this patch introduces functional tests to check
how the condition is updated when Cinder is added as a glance backend.
The test shows how the CinderCondition defined in the glance-operator
transitions from False to True as long as a (default) Cinder CR (with
an arbitrary name) is deployed.

Related: https://issues.redhat.com/browse/OSPRH-14637

[1] https://github.com/openstack-k8s-operators/glance-operator/pull/702

Signed-off-by: Francesco Pantano <fpantano@redhat.com>