Enable notifications for Glance
This patch adds the logic of processing a transportURL CR/secret.
It has been a gap so far compared to the other operators. However,
because Glance do not need RabbitMq to properly work, if a
rabbitmqClusterName is not passed (resulting to an empty string),
the deployment moves forward in any case.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>