Add the ability to deploy httpd+proxypass or wsgi mode
The goal of this patch is to handle minor updates when we move from an
older image that does not support wsgi to a newer image that must be
deployed using the wsgi approach.
This patch adds the ability to drive the Glance deployment through an
annotation that determines the deployment model for a particular image.

Jira: https://issues.redhat.com/browse/OSPRH-16261

Signed-off-by: Francesco Pantano <fpantano@redhat.com>