Fix sporadic failure in serviceOverride envTest
There is an intermittent (not very frequent) issue with the LoadBalancer test
used to check serviceOverrides.
The test is checking for the ReadyCondition to be True, but it might happen
that after the Service is created, the reconciliation loop is still in progress.
This is reinforced by the error output that shows several "Unknown" conditions,
which indicates that the glanceAPI controller is still reconciling.
The issue can be solved in two ways:
1. Ensure that the condition associated to the Service creation is True, and
   mark the test as successful
2. Extend the test and check intermediate conditions before moving forward

We fix the immediate failure by following option 1, and we will extend
the test in a follow up.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>