Simplify apiDeployment
We don't need to pass a memcached instance in the main controller and
let it flow to the reconciliation loop. The only required information
that let us build the apiSpec is instance.Spec.Memcached, which is
known beforehand. In addition, we do not unblock the reconciliation
loop is that field is either not present or it doesn't resolve any
memcached CR in the same namespace.
This patch simplifies how we pass this info across the main controller
functions.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>